# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Social Hub API
  version: "0.1.0"
  description: >
    Unified Social Media Automation & Customer Messaging Platform.
    Auth uses an HttpOnly session cookie (see /api/auth/login) — there is no
    bearer token for the dashboard's own API; webhook endpoints instead use
    per-platform signature verification (see INTEGRATIONS.md).
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - cookieAuth: []
paths:
...
  /api/settings:
    get:
      summary: Get organization settings (Section 48) — business info, AI, automation, notifications, plus summaries
      responses:
        "200": { description: Full settings snapshot for the caller's organization }
    patch:
      summary: Update organization settings (Admin only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                contactEmail: { type: string, format: email, nullable: true }
                contactPhone: { type: string, nullable: true }
                timezone: { type: string }
                automationsEnabled: { type: boolean, description: "Org-wide automation kill switch" }
                aiSystemPromptExtra: { type: string, nullable: true, maxLength: 2000 }
                defaultAiMode: { type: string, enum: [HUMAN, AI, HYBRID] }
                mutedNotificationTypes:
                  type: array
                  items: { type: string, enum: [NEW_MESSAGE, NEW_CUSTOMER, CONVERSATION_ASSIGNED, AI_ESCALATION, INTEGRATION_ERROR, AUTOMATION_FAILED] }
      responses:
        "200": { description: Updated }
        "403": { description: Admin only }
...
  /api/webhooks/whatsapp:
    get:
      summary: Meta webhook verification handshake (hub.challenge)
      security: []
      responses:
        "200": { description: Echoes hub.challenge }
        "403": { description: Invalid verify token }
    post:
      summary: Inbound WhatsApp Cloud API webhook event
      security: []
      description: >
        Verified via the X-Hub-Signature-256 header (HMAC-SHA256 over the
        raw body using META_APP_SECRET). Deduplicated by (platform, a
        derived externalEventId) against the WebhookEvent table.
      responses:
        "200": { description: "{ received: true, processed, duplicates }" }
        "401": { description: Invalid signature }
  /api/webhooks/messenger:
    get: { summary: Meta webhook verification handshake, security: [], responses: { "200": { description: OK } } }
    post: { summary: Inbound Messenger webhook event, security: [], responses: { "200": { description: OK }, "401": { description: Invalid signature } } }
  /api/webhooks/instagram:
    get: { summary: Meta webhook verification handshake, security: [], responses: { "200": { description: OK } } }
    post: { summary: Inbound Instagram webhook event, security: [], responses: { "200": { description: OK }, "401": { description: Invalid signature } } }
  /api/webhooks/tiktok:
    get: { summary: TikTok webhook verification handshake, security: [], responses: { "200": { description: OK } } }
    post:
      summary: Inbound TikTok webhook event
      security: []
      description: >
        Implemented generically (HMAC signature check against
        TIKTOK_WEBHOOK_SECRET); payload parsing is intentionally left unfilled
        until Business Messaging API access is approved for the connected
        account — see INTEGRATIONS.md.
      responses:
        "200": { description: OK }
        "401": { description: Invalid signature }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: social_hub_session
...